+++ src/constants/Router.tsx
@@ -6,7 +6,7 @@ import Login from '@/features/user/Login'
 import Register from '@/features/user/Register'
 import User from '@/features/user/User'
 import ProductDetail from '@/features/productDetail/ProductDetail'
-import ModelFitting from '@/features/productDetail/ModelFitting'
+import ModelFitting from '@/features/product/ModelFitting'
 import Cart from '@/features/order/Cart'
 import Order from '@/features/order/Order'
 import OrderHistoryPage from '@/features/order/OrderHistoryPage'
