+++ src/features/productDetail/components/SizeSelector.tsx
@@ -1,0 +1,41 @@
+import { Select, SoldOut } from '@/features/productDetail/detailStyles'
+import type { SizeInfo } from '@/features/productDetail/types'
+
+interface SizeSelectorProps {
+  selectedSize: string
+  setSelectedSize: (size: string) => void
+  availableSizes: SizeInfo[]
+  isSoldOut: boolean
+}
+
+export const SizeSelector = ({
+  selectedSize,
+  setSelectedSize,
+  availableSizes,
+  isSoldOut,
+}: SizeSelectorProps) => {
+  const getStockDisplay = (quantity: number) => {
+    if (quantity <= 10) {
+      return ` (재고: ${quantity})`
+    }
+    return ''
+  }
+
+  if (isSoldOut) {
+    return <SoldOut>Sold Out</SoldOut>
+  }
+
+  return (
+    <Select
+      value={selectedSize}
+      onChange={(e) => setSelectedSize(e.target.value)}
+    >
+      {availableSizes.map((sizeInfo) => (
+        <option key={sizeInfo.size} value={sizeInfo.size}>
+          {sizeInfo.size}
+          {getStockDisplay(sizeInfo.quantity)}
+        </option>
+      ))}
+    </Select>
+  )
+}
